+++ contracts/vm/ops/math/OpAdd.sol
@@ -9,9 +9,13 @@ import "../../../array/LibUint256Array.sol";
 library OpAdd {
     using LibStackTop for StackTop;
 
-    function add(uint operand_, StackTop stackTop_) internal pure returns (StackTop stackTopAfter_) {
+    function add(uint256 operand_, StackTop stackTop_)
+        internal
+        pure
+        returns (StackTop stackTopAfter_)
+    {
         StackTop location_ = stackTop_.down(operand_);
-        uint accumulator_ = location_.peekUp();
+        uint256 accumulator_ = location_.peekUp();
         stackTopAfter_ = location_.up();
         for (StackTop i_ = stackTopAfter_; i_.lt(stackTop_); i_ = i_.up()) {
             accumulator_ += i_.peekUp();
