+++ contracts/vm/ops/math/OpExp.sol
@@ -14,10 +14,10 @@ library OpExp {
         returns (StackTop stackTopAfter_)
     {
         StackTop location_ = stackTop_.down(operand_);
-        uint accumulator_ = location_.peekUp();
+        uint256 accumulator_ = location_.peekUp();
         stackTopAfter_ = location_.up();
         for (StackTop i_ = stackTopAfter_; i_.lt(stackTop_); i_ = i_.up()) {
-            accumulator_ = accumulator_ ** i_.peekUp();
+            accumulator_ = accumulator_**i_.peekUp();
         }
         location_.set(accumulator_);
     }
