+++ contracts/vm/ops/math/OpMin.sol
@@ -14,10 +14,10 @@ library OpMin {
         returns (StackTop stackTopAfter_)
     {
         StackTop location_ = stackTop_.down(operand_);
-        uint accumulator_ = location_.peekUp();
+        uint256 accumulator_ = location_.peekUp();
         stackTopAfter_ = location_.up();
         for (StackTop i_ = stackTopAfter_; i_.lt(stackTop_); i_ = i_.up()) {
-            uint item_ = i_.peekUp();
+            uint256 item_ = i_.peekUp();
             if (item_ < accumulator_) {
                 accumulator_ = item_;
             }
