+++ contracts/vm/ops/tier/OpITierV2ReportTimeForTier.sol
@@ -8,7 +8,7 @@ import "../../LibStackTop.sol";
 /// @notice Exposes `ITierV2.reportTimeForTier` as an opcode.
 library OpITierV2ReportTimeForTier {
     using LibStackTop for StackTop;
-    using LibStackTop for uint[];
+    using LibStackTop for uint256[];
 
     function stackPops(uint256 operand_)
         internal
@@ -26,10 +26,20 @@ library OpITierV2ReportTimeForTier {
         view
         returns (StackTop)
     {
-        (uint tier_, uint[] memory context_) = stackTop_.list(operand_);
-        (StackTop location_, StackTop stackTopAfter_, uint tierContract_, uint account_) = context_.asStackTop().popAndPeek();
-        location_.set(ITierV2(address(uint160(tierContract_)))
-            .reportTimeForTier(address(uint160(account_)), tier_, context_));
+        (uint256 tier_, uint256[] memory context_) = stackTop_.list(operand_);
+        (
+            StackTop location_,
+            StackTop stackTopAfter_,
+            uint256 tierContract_,
+            uint256 account_
+        ) = context_.asStackTop().popAndPeek();
+        location_.set(
+            ITierV2(address(uint160(tierContract_))).reportTimeForTier(
+                address(uint160(account_)),
+                tier_,
+                context_
+            )
+        );
         return stackTopAfter_;
     }
 }
